pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cap-reference"
version = "0.1.0"
description = "Public-ready FastAPI reference project"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
license-files = ["LICENSE", "NOTICE"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Framework :: FastAPI",
]
dependencies = [
  "cap-protocol[server]>=1.0.0,<2.0.0",
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.28.1,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "pydantic-settings>=2.10.1,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4,<9.0.0",
  "ruff>=0.12.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["abel_cap_server*", "abel_cap_client*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
